+++ include/roq/client/collector.h
@@ -17,6 +17,7 @@ class Collector {
   virtual void operator()(const Event<GatewaySettings> &) {}
   virtual void operator()(const Event<GatewayStatus> &) {}
   virtual void operator()(const Event<ExternalLatency> &) {}
+  virtual void operator()(const Event<RateLimitUsage> &) {}
   virtual void operator()(const Event<ReferenceData> &) {}
   virtual void operator()(const Event<MarketStatus> &) {}
   virtual void operator()(const Event<TopOfBook> &) {}
@@ -27,6 +28,7 @@ class Collector {
   virtual void operator()(const Event<CreateOrder> &) {}
   virtual void operator()(const Event<ModifyOrder> &) {}
   virtual void operator()(const Event<CancelOrder> &) {}
+  virtual void operator()(const Event<CancelAllOrders> &) {}
   virtual void operator()(const Event<OrderAck> &) {}
   virtual void operator()(const Event<OrderUpdate> &) {}
   virtual void operator()(const Event<TradeUpdate> &) {}