+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Added
+
+* `RateLimitUsage` to allow strategy to back off when it high-water mark on
+   rate-limiting has been detected
+* `CancelAllOrders` when all orders must be cancelled immediately
+
 ## 0.7.0 &ndash; 2021-04-15
 
 ### Added